Add the staging and OMOP models
This commit has the bulk of the work which has involved transferring a lot of the code in the original R library into dbt models.
Please refer to the dbt docs and DAG for further information.

This can be refined further both from a performance and explainability POV.

There is also room for making the SQL more dialect agnostic by using dbt macros.