Added typing to and refactored the four python scripts. Details in description.
# get_csv_filepaths.py
- Changed from using os path and directory functions to using pathlib's
Path.
- Added type checking.
- Moved some logic to list/dict comprehension. We can expand this again
if you feel it just makes it more difficult to read.

shard_to_seeds.py
- Again changed to using Path.
- Added typing.
- Cast and type return values from duckdb.
- Separated the list creation and csv creation steps.

# generate_vocab_shard.py
- Uses Path.
- Added type hints.
- Changed some of the SQL queries to keep the list processes inside
duckdb. Previously we were requesting information, cleaning it, putting
it into a list. Then adding that list back into a duckdb table. The new
queries should do this within duckdb.

(duckdb build and building without using seeds ran after this change,
and I was able to use this script on the resulting DB. But I think my
data was from a newer version of Synthea so it may be worth double
checking these SQL queries succeed on a verified data set.)

# generate_dbt_yaml.py
- Added type hints.
- Added Parse Cli arguments function to handle passed in arguments.
- Renamed some classes to use CapWords convention.
- Added _ensure_tag() and _ensure_navigable_string() helper functions to
handle/check return objects from BS4. Slightly hacky here as they return
as a _AtMostOneElement type, which is there internal placeholder for
None | Tag | NavigableString. Helper functions ensure this is actually
the correct type at runtime.
- Refactored some functions.
- cdm_html argument is now --source and is optional. It can be a url or
a path. If not provided then the CDM 5.4 html will be downloaded and
used as source.