get_csv_filepaths.py renamed to get_filepaths.py. Script and load_data_duckdb macro now work with parquet as well.
Changes:
- get_csv_filepaths.py now works with csv and parquet files.
    - Renamed to get_filepaths.py.
    - Will only run on directories containing csv OR containing parquet,
not directories containing both.
- Changed macro 'load_data_duckdb'.
    - Now works with csv or parquet files.
    - Parquet files are added as a view so they do not need to be loaded
into the system.
    - CSV function unchanged, but logic simplified slightly.

Additions:
- csv_to_parquet.py script.
    - Converts a folder of csv files to parquet.
    - Casts the necessary columns to the correct data type, in the same
fashion as the load_data_duckdb macro does.
    - By default will run as if converting synthea files, to convert
vocab files pass the -v or --vocab flag.
    - Output dir can be specified. By default will be in the same dir as
the input directory.

Important Note:
- New script is currently pinned at duckdb version to 1.2.2
- DuckDB version 1.3 was released today 🎉, but the csv sniffer has
improved and now correctly detects the timestamps as timestamp with
timezone, which causes errors in some int models that cast to time.
- This issue also affect the loading of the CSV files if using 1.3.
- We will need to either pin to 1.2.2 (pinning dbt-duckdb version
doesn't pin duckdb) OR update the casts to cast to timetz instead of
time (this works) OR change the conversoin script and the macro (or an
earlier staging model to cast to timestamp (if this works with the
postgres versoin.
- I can include the change in this PR or we can open another if that is
better.