Update requirements files, use pyproject.toml, update readme.
This commit adds a pyproject.toml for managing dependencies and dev tool
configurations.

Dev dependancies are updated slightly:
- Ruff replaces black, but is configured to format in almost exactly the
same way.
- The script dependencies such as beautifulsoup are added to the dev
group.
- Codespell is added as a tool, this catches common spelling mistakes
using a dictionary of common errors (so it won't highlight every little
issue in function/variable names etc). There is a list of ignored words
in it's config in the pyproject.toml file, feel free to add to it. Run
it using condespell <path/to/files>.

The requirements files are updated to specify the current versions of
duckdb and the dbt connectors, for those wanting to stick with pip.

The readme file is updated to show uv vs pip instructions.